Addressed NITs and added class constraint to TProtocol
