+++ src/Swift.Runtime/src/Swift/AnyType.cs
@@ -57,9 +57,10 @@ public struct AnyType : ISwiftObject
     /// <summary>
     /// Gets the protocol conformance descriptor for the given type
     /// </summary>
-    /// <typeparam name="U"></typeparam>
+    /// <typeparam name="TProtocol"></typeparam>
     /// <returns></returns>
-    static ProtocolConformanceDescriptor ISwiftObject.GetProtocolConformanceDescriptor<U>()
+    static ProtocolConformanceDescriptor ISwiftObject.GetProtocolConformanceDescriptor<TProtocol>()
+        where TProtocol : class
     {
         // TODO: Implement
         throw new NotImplementedException();
