+++ src/Swift.Runtime/src/Swift/Runtime/ProtocolWitnessTable.cs
@@ -1,62 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 
 namespace Swift.Runtime;
-
-public readonly struct ProtocolConformanceDescriptor : IEquatable<ProtocolConformanceDescriptor>
-{
-    private readonly IntPtr _handle;
-
-    private ProtocolConformanceDescriptor(IntPtr handle)
-    {
-        _handle = handle;
-    }
-
-    public readonly static ProtocolConformanceDescriptor Zero = default;
-
-    public bool IsValid => _handle != IntPtr.Zero;
-
-    public bool Equals(ProtocolConformanceDescriptor other)
-    {
-        return _handle == other._handle;
-    }
-
-    internal static bool TryGetProtocolConformanceDescriptor<T, U>([NotNullWhen(true)] out ProtocolConformanceDescriptor? result)
-    {
-        var type = typeof(T);
-
-        if (typeof(ISwiftObject).IsAssignableFrom(type))
-        {
-            var helperType = typeof(ProtocolConformanceDescriptorHelper<,>).MakeGenericType(typeof(T), typeof(U));
-            result = (ProtocolConformanceDescriptor)helperType.GetMethod("GetProtocolConformanceDescriptor")!.Invoke(null, null)!;
-            return true;
-        }
-
-        result = null;
-        return false;
-    }
-
-    public static ProtocolConformanceDescriptor LoadProtocolConformanceDescriptor(string libraryPath, string symbolName)
-    {
-        var libraryHandle = NativeLibrary.Load(libraryPath);
-
-        if (libraryHandle == IntPtr.Zero)
-        {
-            throw new SwiftRuntimeException($"Unable to load library: {libraryPath}");
-
-        }
-
-        if (NativeLibrary.TryGetExport(libraryHandle, symbolName, out var handle))
-        {
-            return new ProtocolConformanceDescriptor(handle);
-        }
-
-        throw new SwiftRuntimeException($"Unable to find symbol: {symbolName} in library: {libraryPath}");
-    }
-}
 
 public readonly struct ProtocolWitnessTable : IEquatable<ProtocolWitnessTable>
 {
@@ -76,20 +23,20 @@ public readonly struct ProtocolWitnessTable : IEquatable<ProtocolWitnessTable>
         return _handle == other._handle;
     }
 
-    public ProtocolWitnessTable GetProtocolWitnessTable<T, U>()
+    public static ProtocolWitnessTable GetProtocolWitnessTable<TType, TProtocol>()
+        where TProtocol : class
     {
-        var metadata = TypeMetadata.GetTypeMetadataOrThrow<T>();
-        if (!ProtocolConformanceDescriptor.TryGetProtocolConformanceDescriptor<T, U>(out var conformanceDescriptor))
+        var metadata = TypeMetadata.GetTypeMetadataOrThrow<TType>();
+        if (!ProtocolConformanceDescriptor.TryGetProtocolConformanceDescriptor<TType, TProtocol>(out var conformanceDescriptor))
         {
-            throw new SwiftRuntimeException($"Unable to get protocol conformance descriptor for {typeof(T)} and {typeof(U)}");
+            throw new SwiftRuntimeException($"Unable to get protocol conformance descriptor for {typeof(TType)} and {typeof(TProtocol)}");
         }
 
         return GetProtocolWitnessTable(conformanceDescriptor.Value, metadata);
     }
 
-    ProtocolWitnessTable GetProtocolWitnessTable(ProtocolConformanceDescriptor conformanceDescriptor, TypeMetadata typeMetadata)
+    static ProtocolWitnessTable GetProtocolWitnessTable(ProtocolConformanceDescriptor conformanceDescriptor, TypeMetadata typeMetadata)
     {
-
         var witnessTable = swift_getWitnessTable(conformanceDescriptor, typeMetadata, IntPtr.Zero);
         return witnessTable;
     }
